requirements.txt
python-binance==1.0.19
pandas>=2.0.0
pandas-ta>=0.3.14b
python-dotenv>=1.0.0
colorama>=0.4.6
tabulate>=0.9.0
flask>=3.0.0
